+++ test/unit/vagrant/util/downloader_test.rb
@@ -22,7 +22,9 @@ describe Vagrant::Util::Downloader do
 
   describe "#download!" do
     let(:curl_options) {
-      ["--fail", "--location", "--max-redirs", "10", "--user-agent", described_class::USER_AGENT, "--output", destination, source, {}]
+      ["--fail", "--location", "--max-redirs", "10",
+       "--user-agent", described_class::USER_AGENT,
+       "--output", destination, source, {}]
     }
 
     context "with a good exit status" do
@@ -50,8 +52,25 @@ describe Vagrant::Util::Downloader do
       end
     end
 
-    context "with a UI" do
-      pending "tests for a UI"
+    context "with a username and password" do
+      it "downloads the file with the proper flags" do
+        original_source = source
+        source  = "http://foo:bar@baz.com/box.box"
+        subject = described_class.new(source, destination)
+
+        i = curl_options.index(original_source)
+        curl_options[i] = "http://baz.com/box.box"
+
+        i = curl_options.index("--output")
+        curl_options.insert(i, "foo:bar")
+        curl_options.insert(i, "-u")
+
+        Vagrant::Util::Subprocess.should_receive(:execute).
+          with("curl", *curl_options).
+          and_return(subprocess_result)
+
+        expect(subject.download!).to be_true
+      end
     end
   end